+++ plugins/org.chromium.sdk/src/org/chromium/sdk/internal/liveeditprotocol/LiveEditResult.java
@@ -20,7 +20,7 @@ public interface LiveEditResult {
 
   @JsonType
   interface TextualDiff {
-    long[] chunks();
+    int[] chunks();
   }
 
   @JsonType
